Populating Crew functions and procedures
The crewmember section is gradually coming together, which is what this commit is all about.  Lots to do, but this is a core section of the game and the ship/crew generator.

This effectively consolidates 4 seperate scattered files of pascal code into one; so much nicer.  Don't expect it to work though, still plenty missing.