Planets and Planet items
The majority of this code is implementing the functionality behind the item system used for populating a planetary cache.  Some of the code for planet bot support has also been added, but this needs more work and testing.
Further support functions used by planets has also been added.  Most of the code concerned is taking quite some time, as I'm trying to understand exactly what some of it is doing.  In many cases, this has resulted in massive amounts of simplification of what was there originally.