Yes!  Planet Renderer Working
I know it seems to be a simple thing, but I had no idea how the IS code was achieving what it was with the 2D surface to 2.5D sphere mapping.  There were also functional problems between the languages.

Eventually, I found a SENSIBLE equation that would perform the conversion.  And that is where more pain began, when I discovered that Python's maths library square root function cannot handle negative numbers.  This has led to more having to be done to get around that sensibly, as equation is relies on unit circle math.

Consider this to be the gentle rewrite of the IS rendering engine for Planets.  It has been very educational, and I assure you that the apparent simplicity of the resulting code is inversely proportional to the pain endured.