Planets display and render.
Renderer for planets in place.  This has required a lot of work, given they are also procedurally generated.  Orbit mechanics and rotation + eclipse and sunlight effects will be added later; dynamic resizing is also required.
Assign the class IronSeed to a variable, call the variable with the function .main_loop() to test the intro as it currently stands.  Needs lots of work, but it's getting there.