Mars for intro added, lights for technology on surface added.
A version of mars is now added, and is woirking in the intro.  I am on the lookout for a better seed, though, so that part will change later.  Technology related spots now appear on a planet, and should look pretty nice once the eclipse and brightness effects are added.