Adding more attack sequence code
The sequence will work for the screen being typed up and shrunk into the monitor area of the Alien ship.  However, a hardware layer bug is causing me some grief, and I need to find a workaround.  (If I have to, I might need to make a (1,1,1) pixel using a pixel array at position (0,0) on the shrunk ship image.  This will stop the transparent black being culled, and the image being rammed to coords (0,0) for the area with non (0,0,0) pixels.  Weird stuff...)