Tidy and consolidate functions into crew class.
After looking at the Ego Synth manipulation code, I decided to move some of the functionality into the crew class proper.  This will make changes simpler in the future as it is now more OO than it was before.  Things still nee to be done, but that requires more research on a replacement for the colour palette cycling code found in the Pascal version; this is what is delaying me.