Zooming Works, Map Correct Size, Probots do thing.
The map now displays correctly, as does the zoomed view.  Using the zoom buttons will zoom the view (needs more work to zoom to centre of selected view.)
Buttons, for the most part, are functional.  The Probot logic is operating in a simplistic manner, and drawing for the probot position is working.  Tidying of the movement algo required, and simple things like pausing during certain stages.  Will move onto scan result code once these are tidied.