Formatting + IronSeed time support + multicolour fade
Adjusted formatting for readability.  I've added a specific IronSeed date/time system, with supporting formatting functions, for use in the game displays + Date/Time logic.
The fade function has been extended and now supports fade to different colours, this allows fade to red as is used in the intro and during game over in battle.