Add cargo fit check function
Needed to be able to determine if an object would fit in the ship's cargo hold before attempting to do so.  Exists because it should be common sense someone checks to make sure something will fit before they attempt to ram it ino the cargo hold.
(This is a dark techological future, not idiocracy.)