Enhanced all timers with stopwatch.  Added more graphics functions.
The Planet Scanner now uses the stopwatch, which allows us to have "realtime" probot actions independent of the framerate.  The targetting reticule now works correctly on the zoom monitor, and scales based on the zoom (see drawZoomHUD).  The green text now has its own function (prepareGreenTextFrames), which prepares the "text" textures for drawing this in the zoom window.  It's a little fast, so needs more work.

Refactored much of the probot logic to use the stopwatch, and also went through testing to ensure logic applies correctly.  Still need to test with actual anomalies and data gathering.  Anomalies need more work to add Artifact support, but that is a fairly involved addition to the item loader.  Data gathering is a factor of planet analysis, and needs work on the interpreting the pixels of the planet texture side of things.