Increased planet scanning functionality, added more to HUD draw function
Scanning support has been added in a fuller manner to the planet scanner functions, including a test function to see if scanning is "complete" for the planet, and therefore all scan types.  This unlocks the anomaly rerieval function which is now locked until a planet scan is fully done.

The scanner reset funtion has been added to simplify the exit process, so the scanner is at more of a "blank slate" stage for a new planet.

The probot monitors now display the green processing text when performing an analysis, and also have a red dot for where the probot is in relation to the main landscape display.  This needs a bit more work to ensure the probot monitor overlay graphic display is centred correctly.

We are nearing the stage where we can actually work on the data collection system properly!