Data collection and probot related functions finished.
The data collection side of the Planet Scanner is now finished.  Anomaly functionality has been added and partially tested, but can be considered almost finished; random name artifacts are next on the list, but only need to be plugged into the framework present.  Data display of progress towards planet environment data collection is finished.  Probot functionality has been tested and appears to be finished, perhaps some timings will be adjusted later.
Display panels, scaling adjustments and input related functions remain as part of the next batch of additions and alterations.  The data graphs and text readouts are next to finish, and preliminary function outlines have been created for them.
There are a variety of todos regarding handling event flags, logs and restless natives and hostile planet atmospheres.  However, they need to be  worked on first in relation to another section of the game before we come back here.