Analysis cross adjustment
Adjusted the position of the red cross used during the probot analysis phase, this was to bring it in line with the actual centre where a red ot is drawn.  This will never be 100% perfect due to the way in which scaling is not pixel perfect when doing the float to int conversion.