Added subsetList, return subsets of a list based on index start and end
A helper function which could return a subset of a list was required, this one will take any list and return a new subset list containing the elements inclusive the index start and end values.  This should be vry useful when handling the text drawing later, as we need to be able to scroll up and down a list, and creating a subset should reduce overall calls.