Draw small panel Data summary text
Supporting functions and changes have been added to allow he text summaries to be drawn in the small data panel area.  Very mucha  work in proress, but it is functional.  Now need to followup with cosmetic tidying, and then the pie-graphs for the main window + multi-coloured text lines with %'s of common compounds.  After all that, custom anomaly naming and he anomaly collection screen needs finishing.