Complete Scan windows now displaying textures + text correctly.
The scanner window is now devoid of artefacts from the underlying texture.  The text windows are displaying their readouts correctly for the sub-windows.
The pie graph and common compound text are next.  Following this, or during, will be the mini-biosphere summary window text.  [Anomalies, custom artifacts, artifacts still remain to support in various areas.]