Tidy + use stopwatch more + add tank lights for encode scene
I've been reformating the white space so things are easier for me to read.  You'll notice this as I revisit some code sections.  The tank lights are in place; not absolutely perfect, but this is a precision issue with scaling.  The encode sequence is now using the stopwatch system for timing.  The text has been improved by blanking the screen before each draw.