Tidied IronSeed Logo
The original menu graphic looks like an eraser has sneezed on the logo.  This paritucular effort fills in the missing pixels, in many places, to make it more decent.  This is not perfect, but much better than before.  The lower menu items still need work, and some colours need to be tweaked in the logo in some spots.