Cube Prep work
The Cube memory structure is now created, and the side panels for the cube are linked and populated.  A load routine retrieves all the button position and return data from IronPy_CubeFacets.tab.  Mouse check routine is in place for the cube.

ToDo:  The button tab file, selecting facets, animating the cube rotating [the way this is constructed the cube will now be able to spin as if it is a real object.]  Logic for interpretng the buttons returns touches on many systems, so that will take far longer to get right.  When the buttons are working for already created areas, I'll be able to move away from chain-loading for testing, and be able to use the modules directly.
The planet drawing routine could probably be called relatively soon, but it is not high priority.