Cube Facet Locations and button sizes added
That took a while, the list now contains the x/y coordinates of the buttons, as well as their pixel height and width.  Return codes and button numbers were already present.
This now allows me to move on to the rendering of the cube phase, and then making each individual function accessible + working.  I'll concentrate on drawing the cube and spinning it first, and then pre-exising functions will be linked in.