Added cube drawing, facet change states and button support
The command deck now draws the command cube.  In its primative ersion you can choose the facets using the normal buttons, and then click on the cube buttons.  Some areas are now accessible via state changes, and these are now adjusted accordingly.  It is now possible to perform more extensive testing, and this implements a major part of the core interface overall.  Pretty effects are back-burner.