Crew Status outline class
The crew status button has been enabled, and the initial crew status class is now available.  This allows entry into the crew status area via the command deck crew status button, and displays the raw graphic + plays music for this area.  The exit button is functional.  With outline functions, their presence allows for more testing, and the rapid addition of features as time goes on.