prepareAnimationArray and resizeGraphicArray helper functions added.
To improve overall efficiency in coding, I've added two functions:
prepareAnimationArray - take a texture, and a tuple which defines rows, columns, sarting position, and dimensions of individual frames.  This function essentially extracts frames from a texture and puts them into an array.

resizeGraphicArray - resize all the frames in an array to the current resolution.