Added rotatingLine, textLinesWithin and charactersWithin helper functions
The rotatingLine function returns a character based on a 1 to 8 range, creating a unix pseudo text clock.

The textLinesWithin function offers back the amount of usable text lines based on the font offset, and pixels between the top and bottom parameters.

The charactersWithin function provides the amount of characters possible within a pixel limit defined by start and end variables.  It'll need some more work, but is quite usable right now.