Update for initialization screen support and demo
The initial version of the initialization screen is functional, this provides a replacement of the first two load routines as a demonstration.  The pygame routines have also been moved from the global constants, to fix a bug revealed by Otzen, when multiple initializations of pygame occur.  This brings the initialization screen into place, and allows it to be fully tested while the remaining loading logic is adjusted to support it.  Not all initialization screen logic is fully implemented yet, but this is a good start.