cmath introduced for texture to sphere function
cmath is now imported, as it supports negative numbers for square roots.  This reduces some of the code for the cSphereMap function, and speeds things up slightly; which is highly welcome in this case.
The old sphereMap function still exists, but cSphereMap is preferred.
Formatting in the helper_functions.py file has been improved a little to remove white-space.
planets.py now uses cSphereMap.