Add drawSprayFrame function
drawSprayFrame is a highly inefficient, but working, function which takes a pygame surface, an (x,y) central point to draw a line from, and the pixel increment we are at.  It draws each pixel up to the increment number, and then a line in the same colour as the target pixel from a central point.  The result is returned as a pygame surface.

In IronSeed, when a player first sees the command deck, the cube on the command panel is drawn from left to right, using a per-pixel laser from a central point.  This approximates the graphical look, but it needs serious work to speed it up.