IronSeedTime is now printing on the command deck.  Broke out  first cube draw into function drawCubeFirstTime()
The date and time is now displaying on the command deck, and seems to follow the original 5 second increment requirement.

The drawing of the first cube image has now been broken out into its own function, drawCubeFirstTime, mainly for tidiness.

Note:  The cube drawing is working, so I'm not too concerned about the drawing speed yet, as speed can be looked at later; although, it is horrendous.  The main thing is to get it working first.