Adding Descriptions for buttons and cube facits
The command panel has a description field above the cube, which displays buttons names, the cube side button pressed, and also the same for mouseover.  It is a primative form of tooltip.

The names added to this data file are those descriptions, and should allow simpler modding in the future.