Added support for printing desciprions on command deck description mini window.
The command panel has a description field above the cube, which displays button names, the cube side button pressed, and also the same on mouseover.  It is a primitive form of tool-tip.

class Cubeside has been updated to store and return descriptions for individual buttons and the facit itself.  An update function has been added, in case we need to update a facit name dynamically.

The class cube file loading routine now loads the additional description data, and populates the correct areas accordingly.

In the CommandDeck class, the interaction function will now update the descriptions on mouse-click on a valid button; no mouseover yet.

A new font entry was added to globals and populated in the main ironseed class.  The commanddeck class will now render descriptions using this double-sized font during the rendering phase.