Precalculating sphere mappings and apply cmath
Prior to this change the sphere mapping was calculated afresh for each planet frame.   However, now this is calculated once by populating a dictionary with the coordinates of each transformed pixel location keyed by (x,y).  When game settings for resolution are implemented, beyond the configuration file, these will be updated on in-game resolution change.

SphereMap itself is now using cMath, for added speed.

Further optimisation of the planet colouring stage is planned, probably by storing one pass of the heightmap, and referring to it for rendering; regenerating on planet state change.