Speedup planet rendering
The planet rendering code was made according to the original style of coding present in the original IronSeed, which was working in a computationally intense but memory poor environment.  Compiled Pascal is very fast... Python certainly is not.

So, the planet texture is now precomputed, with colours, to skip repeated processing of this texture for each frame.  This essentially made 0 difference to the rendering speed of the final planet.

Further experimentation always pointed to the same problem, per-pixel access.  After adjusting the planetBitmapToSphere function to a second version, where pre-computation was being fully used, the eclipse functionality was found to be the main bottleneck.

Although the original eclipse code produces a very nice shadowing effect, the rendering time is simply awful.  Compromising by creating an alpha-channel mask, with one single set border region, we can blit this onto the planet frame partway through the process, before the sphere code runs.  This allows the curve to be constructed in real-time, but at the expense of the feathered edge.  Further work will be done on this later, but for now we have sane rendering speeds.