Ensure Prime static is correct size and position
The static for the prime EGO screen is now at the correct size and position for the screen, using more exact coordinates - a result of working on the crew communications module.
Please note, the intro can crash in new builds of python and pygame due to a double-free bug, skip intro if this affects you.